Move the `Search` struct to `search.rs` and the regex util to `utils.rs`
